% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_parameters.R
\name{crop_parameters}
\alias{crop_parameters}
\title{Create a crop_parameters object.}
\usage{
crop_parameters(n_crop, L_ini, L_dev, L_mid, L_late, Kc_ini, Kc_mid,
  Kc_end, rd)
}
\arguments{
\item{n_crop}{Integer scalar. The number of crops in the simulation.}

\item{L_ini}{Numeric vector. The relative length of the inital crop
growth phase.}

\item{L_dev}{Numeric vector. The relative length of the development
crop growth phase.}

\item{L_mid}{Numeric vector.}

\item{L_late}{Numeric vector.}

\item{Kc_ini}{Numeric vector.}

\item{Kc_mid}{Numeric vector.}

\item{Kc_end}{Numeric vector.}

\item{rd}{Numeric vector.}
}
\value{
crop_parameters
}
\description{
Create a crop_parameters object.
}
