% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_parameters.R
\name{farm_parameters}
\alias{farm_parameters}
\title{Create farm_parameters object.}
\usage{
farm_parameters(n_farm, farm_area, field_capacity, wilting_point,
  specific_yield, runoff_coefficient, evaporation_coefficient,
  canal_volume, canal_leakage_rate, irrigation_efficiency)
}
\arguments{
\item{n_farm}{Integer scalar. The number of farms in the simulation.}

\item{farm_area}{Numeric vector. Farm area (m2)}

\item{field_capacity}{Numeric vector. Field capacity (-)}

\item{wilting_point}{Numeric vector. Wilting point (-)}

\item{specific_yield}{Numeric vector. Specific yield (-)}

\item{runoff_coefficient}{Numeric vector. Runoff coefficient (-)}

\item{evaporation_coefficient}{Numeric vector. Evaporation coefficient (-)}

\item{canal_volume}{Numeric vector. Canal volume (m3)}

\item{canal_leakage_rate}{Numeric vector. Canal leakage rate (-)}

\item{irrigation_efficiency}{Numeric vector. Irrigation efficiency (-)}
}
\value{
farm_parameters
}
\description{
Create farm_parameters object.
}
