% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_condition.R
\name{initial_condition}
\alias{initial_condition}
\title{Create initial_condition object.}
\usage{
initial_condition(n_farm, root_zone_depletion, groundwater_depth,
  initial_savings, initial_category)
}
\arguments{
\item{n_farm}{Integer scalar. The number of farms in the model simulation.}

\item{root_zone_depletion}{Numeric vector. Initial root zone depletion.}

\item{groundwater_depth}{Numeric vector. Initial depth to groundwater.}

\item{initial_savings}{Numeric vector. Initial farmer savings.}

\item{initial_category}{Integer vector. Initial category.}
}
\value{
initial_condition
}
\description{
Create initial_condition object.
}
