% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation_schedule.R
\name{irrigation_schedule}
\alias{irrigation_schedule}
\title{Create an irrigation_schedule object.}
\usage{
irrigation_schedule(n_crop, schedule)
}
\arguments{
\item{n_crop}{Integer scalar. The number of crops in the simulation.}

\item{schedule}{List with length equal to \code{n_crop}. Each list
element should be a data.frame with two columns, where the first
column contains the Julian days on which irrigation takes place,
and the second column specifies the depth of irrigation water
to be applied (mm).}
}
\value{
irrigation_schedule
}
\description{
Create an irrigation_schedule object.
}
