% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Create a model object.}
\usage{
model(weather_data, price_data, yield_data, farm_parameters,
  crop_parameters, irrigation_schedule, initial_condition)
}
\arguments{
\item{weather_data}{weather_data. See \code{\link{weather_data}}.}

\item{price_data}{price_data. See \code{\link{price_data}}.}

\item{yield_data}{yield_data. See \code{\link{yield_data}}.}

\item{farm_parameters}{farm_parameters. See \code{\link{farm_parameters}}}

\item{crop_parameters}{crop_parameters. See \code{\link{crop_parameters}}}

\item{irrigation_schedule}{irrigation_schedule. See \code{\link{irrigation_schedule}}}

\item{initial_condition}{initial_condition. See \code{\link{initial_condition}}}
}
\value{
model
}
\description{
Create a model object.
}
