% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_data.R
\name{price_data}
\alias{price_data}
\title{Create a price_data object.}
\usage{
price_data(nitrogen_price, phosphorous_price, potassium_price,
  diesel_price, crop_price, time, ...)
}
\arguments{
\item{nitrogen_price}{Numeric vector, xts. Nitrogen price.}

\item{phosphorous_price}{Numeric vector, xts. Phosphorous price.}

\item{potassium_price}{Numeric vector, xts. Potassium price.}

\item{diesel_price}{Numeric vector, xts. Diesel price.}

\item{crop_price}{data.frame, xts. Crop price.}

\item{time}{POSIX, optional.}

\item{...}{Additional arguments (none)}
}
\value{
price_data
}
\description{
Create a price_data object.
}
