% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data.R
\name{weather_data}
\alias{weather_data}
\title{Create a weather_data object.}
\usage{
weather_data(prec, et0, time)
}
\arguments{
\item{prec}{Numeric vector, xts. Daily precipitation (mm).}

\item{et0}{Numeric vector, xts. Daily reference evapotranspiration (mm).}

\item{time}{POSIX, optional.}
}
\value{
weather_data
}
\description{
Create a weather_data object.
}
