% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_data.R
\name{yield_data}
\alias{yield_data}
\title{Create a yield_data object.}
\usage{
yield_data(yield, time)
}
\arguments{
\item{yield}{Data frame, xts.}

\item{time}{POSIX.}
}
\value{
yield_data
}
\description{
Create a yield_data object.
}
